refactor DELET & PATCH account api
