+++ src/lib/utils/error_dictionary.ts
@@ -708,4 +708,9 @@ export const API_ERRORS = {
     message: "Account has children",
     status: ApiCode.VALIDATION_ERROR,
   } as IErrorDef,
+  VA_ACCOUNT_NOT_FOUND: {
+    code: "VA000040",
+    message: "Account not found",
+    status: ApiCode.VALIDATION_ERROR,
+  } as IErrorDef,
 };
